Refactor middleware to handle form submissions with improved error handling
